Added comments to help explain the stats bar.
It implements Textbox, it can be updated with setText(stats).
The stats string is the result of parsing the HashMap<String,Integer> of
stats name and stat value, which is done in the method getStatsString.

Also, we had to change some of Textbox's methods to protected so that
StatsBar could use them.